+++ b2sdk/v1/sync/__init__.py
@@ -12,5 +12,6 @@ from .encryption_provider import *
 from .file import *
 from .folder import *
 from .folder_parser import *
+from .report import *
 from .scan_policies import *
 from .sync import *