+++ test/unit/sync/test_sync_report.py
@@ -1,6 +1,6 @@
 ######################################################################
 #
-# File: test/unit/v0/test_sync_report.py
+# File: test/unit/sync/test_sync_report.py
 #
 # Copyright 2019 Backblaze Inc. All Rights Reserved.
 #
@@ -9,6 +9,8 @@
 ######################################################################
 
 from unittest.mock import MagicMock
+
+import pytest
 
 from apiver_deps import SyncReport
 
@@ -23,3 +25,30 @@ class TestSyncReport:
         )
         sync_report = SyncReport(stdout, False)
         sync_report.print_completion('transferred: 123.txt')
+
+    @pytest.mark.apiver(to_ver=1)
+    def test_legacy_methods(self):
+        stdout = MagicMock()
+        sync_report = SyncReport(stdout, False)
+
+        assert not sync_report.total_done
+        assert not sync_report.local_done
+        assert 0 == sync_report.total_count
+        assert 0 == sync_report.local_file_count
+
+        sync_report.local_done = True
+        assert sync_report.local_done
+        assert sync_report.total_done
+
+        sync_report.local_file_count = 8
+        assert 8 == sync_report.local_file_count
+        assert 8 == sync_report.total_count
+
+        sync_report.update_local(7)
+        assert 15 == sync_report.total_count
+        assert 15 == sync_report.local_file_count
+
+        sync_report = SyncReport(stdout, False)
+        assert not sync_report.total_done
+        sync_report.end_local()
+        assert sync_report.total_done